Start building things
Getting down some code. I think this might be nice. Since we use Rust
here we can do away with special keycodes and positions and use enums
that can be either a KeyCode or a Function. We can then specify the
layer switching using a Function that alters the state. The same thing
can be done for the leds, build a function that alters the state of the
pins. Think this can be quite nice.