Act upon host sending report
The hosts sends reports to the keyboards about which mode it is in
(which leds should show). Simple acting on that.
Need to figure out better ways to do it. If the leds are already on we
won't see any change. Also, we reset that when we do our own
led-scanning.

I basically need more pins to be able to fix with the leds....