Start working on OnHold and Combo
I want to be able to have more keys on the same key. Layers work for
some things, but being able to just hold keys and make them as like
other keys seems like it could be awesome.
I also want to be able to hit key combos, like KL and have it act like
Enter, or something.
Getting this to work can, in a future keyboard, reduce the need for so
many keys by a lot. Going down to at least 40 seems like no problem at
all if I get this working.