Rewrite state to do ticks
Not done. Not working. But kinda gets the idea. I think I have things
thought out, just need to write the code for it. Using ticks makes it a
lot easier to check for press and release and for how long the button
was pressed.

Laying the groundwork for onhold and combos, basically. A bit worried
about the space needed for this though. Can get quite big I'm afraid so
it might not fit on a 32u4.