Add wrapper for keys
To be able to deal with keys that should trigger directly and keys that
should happen when on hold I have added a wrapper type for keys. This
makes it a lot easier to determine what kind of key it is. The other
option was to do `enum Key {OnHold(Box<Key>, u8, Box<Key>), ...}` which
is a lot more of a mess than adding the wrapper.

All keys are currently `Instant` and the keyboard acts like before the
change.

#2