Not fully working, but specify limit in ms
For some reason key2 is not sent on its own. This is fine if the key is
a key that need some other key to have an effect, like shift. But if you
do `OnHold(KeyCode(W), 100, KeyCode(P))` you kinda want that P to be
sent after 100ms. No matter if some other key is pressed or not. Very
strange why this is not triggering correctly.