2 add onhold (#6)
* Add wrapper for keys

To be able to deal with keys that should trigger directly and keys that
should happen when on hold I have added a wrapper type for keys. This
makes it a lot easier to determine what kind of key it is. The other
option was to do `enum Key {OnHold(Box<Key>, u8, Box<Key>), ...}` which
is a lot more of a mess than adding the wrapper.

All keys are currently `Instant` and the keyboard acts like before the
change.

#2

* Start playing with OnHold

Trying to figure out how to properly trigger the correct thing to
happen. It was a harder nut than I thought to crack.

#2

* Set release time to max on init

The promicro panicked if u8::MAX was set in the constructor. No idea why
it works if you just do it like this. Seem odd.

* Close to working onhold!

Need to fix de-registering of the onhold if pressed. It should be marked
as released but it's not. The states are off.

* Probably fully working onhold!

It works! At least right now. Will need more testing with other keys and
functions. But it seems to be working.

* Not fully working, but specify limit in ms

For some reason key2 is not sent on its own. This is fine if the key is
a key that need some other key to have an effect, like shift. But if you
do `OnHold(KeyCode(W), 100, KeyCode(P))` you kinda want that P to be
sent after 100ms. No matter if some other key is pressed or not. Very
strange why this is not triggering correctly.

* Fully working onhold!

Scmacki dacki baging `<` vs `>` causing trouble for my checks. In some
places I did `time < limit` and others `time > limit`. This made us not
register when key2 was supposed to be sent. We were in the famous
off-by-one-error. By just flipping the GT-check things now work.
Probably should break that check out into a function just to be sure
that I do correct every time. And then it's only one place to change
should I want to.

* Clean up of work

Clean things up so it looks nicer and don't flash the light all the
time.

* Back to basic layout

* Clean up and comment